[Infra][Linux] Move SDK to first-class platform
Replace platform/dynamic_lib with first-class platform/linux targets and
packaging. Update CI and release workflows to publish standard Linux SDK
artifacts.

TEST: Linux x64 SDK release build and checksum verification
