--- platform/dynamic_lib/BUILD.gn
@@ -1,127 +1,0 @@
-# Copyright 2026 The Lynx Authors. All rights reserved.
-# Licensed under the Apache License Version 2.0 that can be found in the
-# LICENSE file in the root directory of this source tree.
-
-import("../../core/Lynx.gni")
-import("../../testing/test.gni")
-
-lynx_core_source_set("lynx_rust_capi") {
-  sources = [ "lynx_rust_capi.cc" ]
-
-  configs = [
-    "../embedder:config",
-    "../embedder:public_api",
-    "../../clay:config",
-  ]
-
-  deps = [ "../../third_party/skia" ]
-
-  include_dirs = [ "../.." ]
-}
-
-config("windowless_embedder_config") {
-  defines = [ "ENABLE_WINDOWLESS" ]
-}
-
-lynx_core_source_set("dynamic_lib_windowless") {
-  sources = [
-    "../embedder/windowless/lynx_ui_renderer_windowless.cc",
-    "../embedder/windowless/lynx_ui_renderer_windowless.h",
-    "../embedder/windowless/lynx_windowless_renderer.cc",
-    "../embedder/windowless/lynx_windowless_renderer_priv.h",
-    "../embedder/windowless/windowless_ui_task_runner_delegate.cc",
-    "../embedder/windowless/windowless_ui_task_runner_delegate.h",
-  ]
-
-  public_configs = [
-    "../embedder:public_api",
-    "../embedder:config",
-    ":windowless_embedder_config",
-  ]
-
-  deps = [
-    "../../clay/lynx_adaptor:resource_loader_embedder",
-    "../../clay/shell/platform/headless:headless",
-    "../../clay/ui:ui",
-  ]
-}
-
-lynx_core_source_set("windowless_embedder") {
-  sources = [
-    "../embedder/frame_timing_listener_impl.cc",
-    "../embedder/frame_timing_listener_impl.h",
-    "../embedder/lynx_env.cc",
-    "../embedder/lynx_group.cc",
-    "../embedder/lynx_group_priv.h",
-    "../embedder/lynx_load_meta.cc",
-    "../embedder/lynx_load_meta_priv.h",
-    "../embedder/lynx_log.cc",
-    "../embedder/lynx_native_view.cc",
-    "../embedder/lynx_runtime_lifecycle_observer_priv.h",
-    "../embedder/lynx_template_bundle.cc",
-    "../embedder/lynx_template_bundle_priv.h",
-    "../embedder/lynx_template_data.cc",
-    "../embedder/lynx_template_data_priv.h",
-    "../embedder/lynx_trace.cc",
-    "../embedder/lynx_update_meta.cc",
-    "../embedder/lynx_update_meta_priv.h",
-    "../embedder/lynx_view.cc",
-    "../embedder/lynx_view_builder.cc",
-    "../embedder/lynx_view_builder_priv.h",
-    "../embedder/lynx_view_client.cc",
-    "../embedder/lynx_view_client_priv.h",
-    "../embedder/lynx_view_clients.cc",
-    "../embedder/lynx_view_clients.h",
-    "../embedder/lynx_view_priv.h",
-    "../embedder/lynx_vsync_monitor.cc",
-    "../embedder/lynx_vsync_monitor_priv.h",
-    "../embedder/memory/lynx_memory.cc",
-  ]
-  if (is_win || is_mac) {
-    sources += [
-      "../embedder/native_view/lynx_texture_view.cc",
-      "../embedder/native_view/lynx_texture_view.h",
-    ]
-  }
-
-  deps = [
-    ":dynamic_lib_windowless",
-    "../../core:lynx_native",
-    "../embedder:public_headers",
-    "../embedder/core:renderer",
-    "../embedder/lynx_service:lynx_service",
-    "../embedder/resource:resource",
-  ]
-  if (enable_napi_binding) {
-    public_deps = [
-      "../../third_party/napi:weak_node_api_host",
-      "../embedder:lynx_runtime_lifecycle_observer",
-      "../embedder/module:module",
-    ]
-  }
-  if (enable_inspector && enable_logbox) {
-    deps += [ "../embedder/lynx_devtool:lynx_devtool_logbox" ]
-  }
-
-  public_configs = [
-    "../embedder:public_api",
-    "../embedder:config",
-    ":windowless_embedder_config",
-  ]
-
-  if (!enable_unittests && !is_android && !is_harmony) {
-    sources += [
-      "../embedder/sysinfo_embedder.cc",
-      "../embedder/vsync_monitor_fallback.cc",
-      "../embedder/vsync_monitor_fallback.h",
-    ]
-  } else {
-    sources += [
-      "../../clay/lynx_adaptor/native_platform_view_mock.cc",
-      "../embedder/lynx_ui_renderer_mock.cc",
-    ]
-  }
-  if (is_win) {
-    sources += [ "../../core/renderer/utils/sys_lynx_env/sys_env_default.cc" ]
-  }
-}
