--- platform/dynamic_lib/macos/BUILD.gn
@@ -1,86 +1,0 @@
-# Copyright 2026 The Lynx Authors. All rights reserved.
-# Licensed under the Apache License Version 2.0 that can be found in the
-# LICENSE file in the root directory of this source tree.
-
-import("../../../core/Lynx.gni")
-
-source_set("macos_common") {
-  sources = [
-    "../../darwin/macos/common/LynxUIRenderer.h",
-    "../../darwin/macos/common/LynxUIRenderer.mm",
-    "../../darwin/macos/common/vsync_monitor_macos.h",
-    "../../darwin/macos/common/vsync_monitor_macos.mm",
-  ]
-
-  configs += [
-    "../../../clay:config",
-    "../../../clay/shell/platform/darwin/common:framework_relative_headers",
-  ]
-
-  deps = [
-    "..:windowless_embedder",
-    "../../../clay/ui:ui",
-    "../../../third_party/rapidjson",
-  ]
-}
-
-shared_library("lynx_clay_shared_library") {
-  output_name = "Lynx_clay"
-
-  deps = [
-    ":macos_common",
-    "..:lynx_rust_capi",
-    "..:windowless_embedder",
-    "../../../base/src:base_static",
-    "../../../clay/lynx_adaptor:lynx_adaptor",
-    "../../../clay/shell/platform/darwin/macos:flutter_framework_source",
-    "../../../clay/shell/platform/embedder:embedder",
-    "../../../core/runtime/common/napi:napi_binding_jsc",
-  ]
-  if (enable_testbench_replay) {
-    deps += [ "../../../core/services/replay" ]
-  }
-  if (enable_napi_binding) {
-    deps += [
-      "../../../core/runtime/common/napi:napi_binding_quickjs",
-      "../../../third_party/napi:env",
-    ]
-  }
-  if (jsengine_type == "v8") {
-    deps += [ "../../../third_party/NativeScript:NativeScript" ]
-    if (enable_napi_binding) {
-      deps += [
-        "../../../core/runtime/common/napi:napi_binding_v8",
-        "../../../third_party/napi:v8",
-      ]
-    }
-  }
-
-  ldflags = [
-    "-Wl,-install_name,@rpath/libLynx_clay.dylib",
-    "-Wl,-exported_symbols_list," +
-        rebase_path("exported_symbols.sym", root_build_dir),
-  ]
-
-  frameworks = [
-    "CoreData.framework",
-    "QuartzCore.framework",
-    "WebKit.framework",
-    "JavaScriptCore.framework",
-  ]
-}
-
-action("package_sdk") {
-  script = "package_sdk.py"
-  outputs = [ "$root_build_dir/lynx_clay_sdk_macos_${target_cpu}.zip" ]
-  args = [
-    rebase_path(outputs[0]),
-    rebase_path(root_build_dir),
-    rebase_path("../../../third_party/icu/flutter/icudtl.dat"),
-  ]
-  deps = [
-    ":lynx_clay_shared_library",
-    "../../darwin/macos:lynx_debug_resources_bundle",
-    "../../darwin/macos:lynx_resources_bundle",
-  ]
-}
