+++ platform/linux/BUILD.gn
@@ -1,0 +1,64 @@
+# Copyright 2026 The Lynx Authors. All rights reserved.
+# Licensed under the Apache License Version 2.0 that can be found in the
+# LICENSE file in the root directory of this source tree.
+
+import("../../core/Lynx.gni")
+
+shared_library("linux") {
+  output_name = "lynx"
+
+  sources = [
+    "../embedder/lynx_ui_renderer_mock.cc",
+    "lynx_platform_fallbacks.cc",
+  ]
+
+  deps = [ ":lynx_linux_group" ]
+
+  libs = [
+    "epoxy",
+    "rt",
+  ]
+}
+
+group("lynx_linux_group") {
+  deps = [
+    "../../base/src:base_static",
+    "../../clay/lynx_adaptor:lynx_adaptor",
+    "../../clay/shell/platform/embedder:embedder",
+    "../embedder:embedder",
+  ]
+  if (enable_napi_binding) {
+    deps += [
+      "../../core/runtime/common/napi:napi_binding_quickjs",
+      "../../third_party/napi:env",
+      "../../third_party/napi:quickjs",
+      "../../third_party/quickjs",
+    ]
+    if (jsengine_type == "v8") {
+      deps += [
+        "../../core/runtime/common/napi:napi_binding_v8",
+        "../../third_party/napi:v8",
+      ]
+    }
+  }
+}
+
+build_kernel_js("lynx_core") {
+  script = "../../tools/js_tools/build.py"
+  platform = "ios"
+  output_path = "$root_build_dir/lynx_core"
+}
+
+action("package_sdk") {
+  script = "./package_sdk.py"
+  outputs = [ "$root_build_dir/lynx_sdk_linux_${target_cpu}.zip" ]
+  args = [
+    rebase_path(outputs[0]),
+    rebase_path(root_build_dir),
+    rebase_path("//third_party/icu/flutter/icudtl.dat"),
+  ]
+  deps = [
+    ":linux",
+    ":lynx_core",
+  ]
+}
