+++ platform/linux/lynx_platform_fallbacks.cc
